requirements.txt
selenium>=4.15.0
python-dotenv>=1.0.0
pytz>=2024.1
webdriver-manager>=4.0.0
requests>=2.31.0